Module pallet_glutton::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

bloat

Increase the block size by including the specified garbage bytes.

initialize_pallet

Initialize the pallet. Should be called once, if no genesis state was provided.

set_block_length

Set how much of the block length should be filled with trash data on each block.

set_compute

Set how much of the remaining ref_time weight should be consumed by on_idle.

set_storage

Set how much of the remaining proof_size weight should be consumed by on_idle.