Crate pallet_glutton

WARNING

DO NOT USE ON VALUE-BEARING CHAINS. THIS PALLET IS ONLY INTENDED FOR TESTING USAGE.

Glutton Pallet

Pallet that consumes ref_time and proof_size of a block. Based on the Compute and Storage parameters the pallet consumes the adequate amount of weight.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_glutton

Constants

MAX_TRASH_DATA_ENTRIES

Max number of entries for the TrashData map.

RESOURCE_HARD_LIMIT

Hard limit for any other resource limit (in units).

VALUE_SIZE

The size of each value in the TrashData storage in bytes.